requirements-docker.txt
# Core Python packages - pinned from working venv
# Generated from working venv on 2025-11-13

# Web Framework & API
fastapi==0.121.1
uvicorn[standard]==0.38.0
httptools==0.7.1
watchfiles==1.1.1
websockets==15.0.1
python-multipart==0.0.20
starlette==0.49.3

# Task Queue
celery==5.5.3
redis==6.4.0
flower==2.0.1
amqp==5.3.1
billiard==4.2.2
kombu==5.5.4
vine==5.1.0

# Database
psycopg2-binary==2.9.11
pgvector==0.4.1
asyncpg==0.30.0

# Supabase Client
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
storage3==2.22.0
postgrest==2.22.0
realtime==2.22.0

# Docling SDK & Dependencies
docling==2.61.2
docling-core==2.51.0
docling-ibm-models==3.10.2
docling-parse==4.7.1

# OCR & Document Processing
easyocr==1.7.2
pytesseract==0.3.13
pypdf==6.2.0
pypdfium2==4.30.0
python-docx==1.2.0
python-pptx==1.0.2

# PDF & Document Libraries
pyclipper==1.3.0.post6
python-bidi==0.6.7
rapidocr==3.4.2

# Machine Learning & AI
torch==2.9.1
torchvision==0.24.1
transformers==4.57.1
tokenizers==0.22.1
huggingface-hub==0.36.0
safetensors==0.6.2
accelerate==1.11.0

# OpenAI & Embeddings
openai==2.5.0
tiktoken==0.12.0
jiter==0.11.1

# Image Processing
Pillow==11.3.0
opencv-python-headless==4.12.0.88
ImageIO==2.37.2
scikit-image==0.25.2
tifffile==2025.10.16

# AWS S3
boto3==1.40.55
botocore==1.40.55
s3transfer==0.14.0

# Google Drive Integration
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0
pyparsing==3.2.1
google-api-core==2.23.0
googleapis-common-protos==1.66.0
proto-plus==1.25.0
protobuf==5.29.3

# Data Processing
numpy==2.2.6
pandas==2.3.3
scipy==1.16.3
openpyxl==3.1.5
xlsxwriter==3.2.9

# Text Processing & NLP
langchain-core==1.0.0
langchain-text-splitters==1.0.0
langsmith==0.4.37
semchunk==2.2.2
tree-sitter==0.23.2
tree-sitter-c==0.23.6
tree-sitter-java==0.23.5
tree-sitter-javascript==0.23.1
tree-sitter-python==0.23.6
tree-sitter-typescript==0.23.2

# Markup & Document Formats
lxml==6.0.2
beautifulsoup4==4.14.2
soupsieve==2.8
markdown-it-py==4.0.0
marko==2.2.1
mdurl==0.1.2

# Configuration & Environment
python-dotenv==1.1.1
pydantic==2.12.3
pydantic_core==2.41.4
pydantic-settings==2.12.0
omegaconf==2.3.0

# HTTP Clients
requests==2.32.5
requests-toolbelt==1.0.0
httpx==0.28.1
httpcore==1.0.9
h2==4.3.0
h11==0.16.0
hpack==4.1.0
hyperframe==6.1.0

# Utilities
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorlog==6.10.1
rich==14.2.0
Pygments==2.19.2
typer==0.19.2
shellingham==1.5.4
tqdm==4.67.1
humanize==4.14.0

# Data Structures & Serialization
attrs==25.4.0
orjson==3.11.3
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1

# Cryptography & Security
cryptography==46.0.3
PyJWT==2.10.1
certifi==2025.10.5
cffi==2.0.0
pycparser==2.23

# System & OS
psutil==7.1.3
filelock==3.20.0
filetype==1.2.0
fsspec==2025.10.0
aiofiles==25.1.0

# Math & Scientific
sympy==1.14.0
mpmath==1.3.0
networkx==3.5
shapely==2.1.2
rtree==1.4.1

# XML & Data Formats
et_xmlfile==2.0.0
latex2mathml==3.78.1
pylatexenc==2.10

# Multiprocessing & Concurrency
dill==0.4.0
multiprocess==0.70.18
mpire==2.10.2

# Testing & Mocking
Faker==38.0.0
polyfactory==2.22.4
pluggy==1.6.0

# Date & Time
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# Monitoring & Prometheus
prometheus_client==0.23.1

# CLI & Terminal
prompt_toolkit==3.0.52
wcwidth==0.2.14

# Type Checking & Validation
typing_extensions==4.15.0
typing-inspection==0.4.2
annotated-types==0.7.0
annotated-doc==0.0.4

# Miscellaneous
anyio==4.11.0
sniffio==1.3.1
idna==3.11
charset-normalizer==3.4.4
urllib3==2.5.0
packaging==25.0
six==1.17.0
colorama==0.4.6
deprecation==2.1.0
distro==1.9.0
Jinja2==3.1.6
MarkupSafe==3.0.3
referencing==0.37.0
rpds-py==0.28.0
PyYAML==6.0.3
tabulate==0.9.0
tenacity==9.1.2
regex==2025.9.18
lazy_loader==0.4
propcache==0.4.1
yarl==1.22.0
multidict==6.7.0
zstandard==0.25.0
ninja==1.13.0
StrEnum==0.4.15
tornado==6.5.2
antlr4-python3-runtime==4.9.3

# YouTube video processing
yt-dlp==2024.12.23
ffmpeg-python==0.2.0
